CI changes
- stop pipeline when a build dev-images stage fails
- better diagnostics: show number of failed workers for a given env (k3d or minikube)
  when a worker fails, then the tests it was meant to run are not included in the junit xml result,
  so it may mislead at first sight (as the build may look quite successful)
- log kubectl port-forward output

+++ tests/ci/registry/build-enterprise-image.sh
@@ -22,4 +22,8 @@ CONTAINER=$(docker container create ${BASE_OPERATOR_IMAGE})
 sed 's/Edition.community/Edition.enterprise/' < mysqloperator/controller/config.py > mysqloperator/controller/config.enterprise.py
 docker container cp mysqloperator/controller/config.enterprise.py ${CONTAINER}:/usr/lib/mysqlsh/python-packages/mysqloperator/controller/config.py
 docker container commit ${CONTAINER} ${ENTERPRISE_OPERATOR_IMAGE}
+if [ $? -ne 0 ]; then
+	echo "cannot build enterprise dev-image ${LOCAL_REGISTRY_ENTERPRISE_OPERATOR_IMAGE}"
+	exit 2
+fi
 docker container rm ${CONTAINER}
